fixed things uo
